Разбивка на модули и доработка фукнции генерации массива объекта
