+++ js/main.js
@@ -1,5 +1,3 @@
-import {createArrayUserDescription} from './demo-data.js';
+import {photoDescriptions} from './mock/demo-data.js';
 
-const photoDescriptionArray = new Array(25).fill(null).map(createArrayUserDescription);
-
-photoDescriptionArray;
+photoDescriptions;
